Move @@tmpnam test to a separate file for isolation
